refactor(orchestrator): extract key/id utility functions into keys.rs
Extract 50 pure tag/key utility functions from main.rs into a new
keys.rs module. These are idempotent ID generators, claim keys,
and tag name formatters used throughout the runtime.

main.rs: 72084 -> 71729 lines (355 fewer)
keys.rs: 405 lines (new)

Also:
- Fix clippy ceiling check (236 warnings, unchanged)
- Bump version to 0.0.331
