+++ v2/scripts/main-rs-line-ceiling.txt
@@ -180,4 +180,4 @@
 # reply wrapper's only remaining caller went with them. Its test duplicated
 # coverage that already lives in ai-model-rust/src/lib.rs, so it moved out
 # rather than being copied again.
-72084
+71729
